# src/main/resources/api-swagger.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: Tracking Service API
  description: >-
    <hr/>
      Provide apis to support upload and retrieve tracking information
    <br/>
     <br/>
    TRACKING SERVICE API
    <br/>
    
    <ul>
      <li>Upload tracking file</li>
      <li>Find latest track</li>
      <li>Get track detail</li>
    <br/>

servers:  
  - url: /tracking

paths:
  '/{userId}/upload':
    post:
      tags:
        - Tracking service
      operationId: uploadFile
      summary: Upload gpx file
      description: >-
        An endpoint that allow users to upload "gpx" file and store mandatory information such as "metadata, waypoint, track" 
      parameters:
        - name: userId
          in: path
          required: true
          description: an existing id in database
          schema:
            type: string
          example: '1'
      requestBody:
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                gpxFile:
                  type: string
                  format: binary
      responses:
        202:
          description: Upload is successfully.
          content:
            text/plain:
              schema:
                type: string
              example: 
                File sample.gpx is successfully uploaded!
        400:
          description: Upload is not successfully.
          content:
            text/plain:
              schema:
                type: string
              examples:
                EmptyFile: 
                  value:
                    Please select a file!
                SaveError: 
                  value:
                    Error while reading file data...

  '/latest':
    get:
      tags:
        - Tracking service
      operationId: getLatestTrack
      summary: Find latest track
      description: >-
        An endpoint to return a list of "Latest track" from our users
      parameters:
        - name: page
          in: query
          required: false
          description: paginator value
          schema:
            type: string
          example: '0'
        - name: size
          in: query
          required: false
          description: paginator value
          schema:
            type: string
          example: 10
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    fileId:
                      type: integer
                    fileName:
                      type: string
                    uploadTime:
                      type: string
                    username:
                      type: string
              example: 
                - fileId: 4
                  fileName: Chris_running
                  uploadTime: '2020-07-14T17:41:33.000+00:00'
                  username: Alice
                - fileId: 3
                  fileName: Bob_trekking
                  uploadTime: '2020-07-14T13:41:33.000+00:00'
                  username: Bob
                  
  '/detail':
    get:
      tags:
        - Tracking service
      operationId: getFileDetail
      summary: Get file detail
      description: >-
        An endpoint to allow users to view details of their gpx file
      parameters:
        - name: id
          in: query
          required: true
          description: a valid fileId
          schema:
            type: string
          example: '1'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  metadata:
                    type: object
                    properties:
                      name:
                        type: string
                      description:
                        type: string
                      author:
                        type: string
                      link:
                        type: object
                        properties:
                          linkHref:
                            type: string
                          linkText:
                            type: string
                      time:
                        type: string
                  waypoints:
                    type: array
                    items:
                      type: object
                      properties:
                        latitude:
                          type: number
                        longitude:
                          type: number
                        name:
                          type: string
                        symbol:
                          type: string
                  track:
                    type: object
                    properties:
                      trackSegmentDTOs:
                        type: array
                        items:
                          type: object
                          properties:
                            trackPoints:
                              type: array
                              items:
                                type: object
                                properties:
                                  latitude:
                                    type: number
                                  longitude:
                                    type: number
                                  elevation:
                                    type: number
                                  time:
                                    type: string
              example: 
                metadata:
                  name: Sahara trekking
                  description: Enter sandman
                  author: BoB
                  link:
                    linkHref: 'https://www.strava.com/'
                    linkText: Strava
                  time: '2020-07-10T09:41:33.000+00:00'
                waypoints:
                  - latitude: 64.4444
                    longitude: 75.5555
                    name: B6
                    symbol: /cdn/B6
                track:
                  trackSegmentDTOs:
                    - trackPoints:
                        - latitude: 62.151515
                          longitude: 63.161616
                          elevation: 61.141414
                          time: '2020-07-10T09:41:33.000+00:00'
        400:
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
              example: 
                Query parameter 'id' is required!
        404:
          description: Not Found
          content:
            text/plain:
              schema:
                type: string
              example: 
                fileId=33 not found.
                